# frontend/src/openapi-codegen/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clochette API",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v2/non_inventoried/": {
      "get": {
        "tags": [
          "non_inventoried"
        ],
        "summary": "Read Non Inventorieds",
        "description": "Retrieve a list of non inventorieds.",
        "operationId": "read_non_inventorieds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NonInventoried"
                  },
                  "type": "array",
                  "title": "Response Read Non Inventorieds"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "non_inventoried"
        ],
        "summary": "Create Non Inventoried",
        "description": "Create a new non inventoried.",
        "operationId": "create_non_inventoried",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NonInventoriedCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoried"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ]
      }
    },
    "/api/v2/non_inventoried/{non_inventoried_id}": {
      "get": {
        "tags": [
          "non_inventoried"
        ],
        "summary": "Read Non Inventoried",
        "description": "Retrieve a non inventoried.",
        "operationId": "read_non_inventoried",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "non_inventoried_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Non Inventoried Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoried"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/non_inventoried_item/": {
      "get": {
        "tags": [
          "non_inventoried_item"
        ],
        "summary": "Read Non Inventoried Items",
        "description": "Retrieve a list of non inventoried items.\n\nQuery parameters:\n    - `trade`: The trade type.\n    - `name`: The non inventoried item name.",
        "operationId": "read_non_inventoried_items",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "trade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TradeType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NonInventoriedItem"
                  },
                  "title": "Response Read Non Inventoried Items"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "non_inventoried_item"
        ],
        "summary": "Create Non Inventoried Item",
        "description": "Create a non inventoried item.",
        "operationId": "create_non_inventoried_item",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NonInventoriedItemCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoriedItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/non_inventoried_item/{non_inventoried_item_id}": {
      "get": {
        "tags": [
          "non_inventoried_item"
        ],
        "summary": "Read Non Inventoried Item",
        "description": "Retrieve a non inventoried item.",
        "operationId": "read_non_inventoried_item",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "non_inventoried_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Non Inventoried Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoriedItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "non_inventoried_item"
        ],
        "summary": "Update Non Inventoried Item",
        "description": "Update a non inventoried item.",
        "operationId": "update_non_inventoried_item",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "non_inventoried_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Non Inventoried Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NonInventoriedItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoriedItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "non_inventoried_item"
        ],
        "summary": "Delete Non Inventoried Item",
        "description": "Delete a non inventoried item.",
        "operationId": "delete_non_inventoried_item",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "staff"
            ]
          }
        ],
        "parameters": [
          {
            "name": "non_inventoried_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Non Inventoried Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonInventoriedItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "example": "Internal server error"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IconName": {
        "type": "string",
        "enum": [
          "Glass",
          "Beer",
          "Food",
          "Soft",
          "Barrel",
          "Misc"
        ],
        "title": "IconName"
      },
...
      "NonInventoried": {
        "properties": {
          "nonInventoriedItemId": {
            "type": "integer",
            "title": "Noninventorieditemid"
          },
          "buyPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyprice"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sellPrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "readOnly": true
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IconName"
              }
            ],
            "readOnly": true
          },
          "trade": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TradeType"
              }
            ],
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "nonInventoriedItemId",
          "id",
          "sellPrice",
          "name",
          "icon",
          "trade"
        ],
        "title": "NonInventoried"
      },
      "NonInventoriedCreate": {
        "properties": {
          "nonInventoriedItemId": {
            "type": "integer",
            "title": "Noninventorieditemid"
          },
          "buyPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyprice"
          },
          "transactionId": {
            "type": "integer",
            "title": "Transactionid"
          }
        },
        "type": "object",
        "required": [
          "nonInventoriedItemId",
          "transactionId"
        ],
        "title": "NonInventoriedCreate"
      },
      "NonInventoriedItem": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "icon": {
            "$ref": "#/components/schemas/IconName"
          },
          "sellPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "trade": {
            "$ref": "#/components/schemas/TradeType"
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon",
          "id",
          "trade"
        ],
        "title": "NonInventoriedItem"
      },
      "NonInventoriedItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "icon": {
            "$ref": "#/components/schemas/IconName"
          },
          "sellPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon"
        ],
        "title": "NonInventoriedItemCreate"
      },
      "NonInventoriedItemUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IconName"
              },
              {
                "type": "null"
              }
            ]
          },
          "sellPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sellprice"
          }
        },
        "type": "object",
        "title": "NonInventoriedItemUpdate"
      },
...
      "TradeType": {
        "type": "string",
        "enum": [
          "purchase",
          "sale"
        ],
        "title": "TradeType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "staff": "Manage stock and create transactions",
              "treasurer": "Manage treasury, i.e. CRUD over transactions",
              "president": "Manage users, i.e. CRUD over users"
            },
            "tokenUrl": "/api/v1/auth/login/"
          }
        }
      }
    }
  }
}